Use a constant amount of CPU work on each request
Looping until wall clock has advanced by 10ms doesn't always result
in 10ms of CPU work because if the thread is suspended (prempted or
stopped to let GC run), the wall clock keeps ticking.

Especially since `Time.now` allocate an object each time, and on 3.2
and older these objects are not write barrier protected and also
do a malloc, so they stress the GC and are likely to trigger it.

To avoid that, it's preferable to precompute a number of iterations
needed to crunch the CPU for ~10ms, and then always do that same amount
of work.